CF: styling home page
