+++ internationalcandyemporium/src/main/java/com/gcu/internationalcandyemporium/Controllers/HomeController.java
@@ -1,16 +1,21 @@
 package com.gcu.internationalcandyemporium.Controllers;
 
+import com.gcu.internationalcandyemporium.Data.ProductsDataService;
 import com.gcu.internationalcandyemporium.Models.ProductModel;
 import com.gcu.internationalcandyemporium.Service.ProductsService;
 
-import java.util.List; // Correct  for List
+import java.util.List;
+
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.ModelAndView;
-
 
 @Controller
 @RequestMapping("/")
@@ -19,30 +24,38 @@ public class HomeController {
     @Autowired
     private ProductsService productsService;
 
+    @Autowired
+    private ProductsDataService productsDataService;
+
     @GetMapping("/")
     public String home(Model model) {
         List<ProductModel> products = productsService.getAllProducts();
-        // Add a logging statement to inspect the products list
-        System.out.println("Products: " + products);
         model.addAttribute("products", products);
+        model.addAttribute("product", new ProductModel(null, null, null, 0, 0));
         return "home";
     }
-    
 
+    @PostMapping("/addProduct")
+    public String addProduct(@ModelAttribute @Valid ProductModel product, BindingResult result, Model model) {
+        if (result.hasErrors()) {
+            result.getAllErrors().forEach(error -> System.out.println(error.getDefaultMessage()));
+            return "home";
+        }
+        
+        productsDataService.create(product);
+        return "redirect:/";
+    }
+    
     //future profile
     @GetMapping("/profile")
-    public ModelAndView profile() {
-        // Logic to handle profile view
-        ModelAndView modelAndView = new ModelAndView("profile");
-        return modelAndView;
+    public String profile() {
+        return "profile";
     }
 
     //future cart
     @GetMapping("/cart")
-    public ModelAndView viewCart() {
-        // Logic to handle cart view
-        ModelAndView modelAndView = new ModelAndView("cart");
-        return modelAndView;
+    public String viewCart() {
+        return "cart";
     }
 
     @GetMapping("/login")
@@ -50,16 +63,11 @@ public class HomeController {
         return "login";
     }
 
-    //future logout
-    @GetMapping("/logout")
-    public ModelAndView logOut() {
-        // Logic to handle logout
-        ModelAndView modelAndView = new ModelAndView("logout");
-        return modelAndView;
-    }
+    // @GetMapping("/logout")
+    // public String logOut() {
+    //     model.addAttribute("title", "Login Form");
+    //     return "redirect:/login";
+    // }
 
-    @GetMapping("/admin")
-    public String admin() {
-        return "admin";
-    }
 }
+
