+++ internationalcandyemporium/src/main/resources/static/style.css
@@ -102,6 +102,11 @@ input {
     height: 50px;
     font-size: x-large;
     border-radius: 10%;
+    cursor: pointer;
+}
+
+.button:hover {
+    background-color: #52061e;
 }
 
 p {
@@ -109,20 +114,22 @@ p {
 }
 
 .add-to-cart {
+    margin-top: 5px;
+    padding-bottom: 6px;
+    padding-left: 7px;
     font-size: 24px; /* Adjust the size of the '+' sign */
     text-align: center;
-    width: 50px; /* Width of the button */
-    height: 50px; /* Height of the button */
+    width: 45px; /* Width of the button */
+    height: 45px; /* Height of the button */
     border: 2px solid white; 
     color: white;/* Border of the button */
-    background-color: rgba(255, 255, 255, 0.5); /* White background with 50% opacity */
+    background-color: #819884; /* White background with 50% opacity */
     line-height: 50%; /* Aligns the '+' vertically */
     border-radius: 50%; /* Makes the button circular */
     cursor: pointer; /* Changes cursor to pointer on hover */
   }
 
   .add-to-cart-container {
-    /* position: absolute; */
     display: flex;
     justify-content: flex-end;
     margin-bottom: 0;
@@ -136,4 +143,53 @@ p {
   .quantity-select {
     width: 50px;
   }
+
+  /* The Modal (background) */
+.modal {
+    display: none; /* Hidden by default */
+    position: fixed; /* Stay in place */
+    z-index: 1; /* Sit on top */
+    left: 0;
+    top: 0;
+    width: 100%; /* Full width */
+    height: 100%; /* Full height */
+    overflow: auto; /* Enable scroll if needed */
+    background-color: rgb(0,0,0); /* Fallback color */
+    background-color: rgba(0,0,0,0.4);
+  }
   
+  /* Modal Content */
+  .modal-content {
+    background-color: #bdd3c0;
+    margin: 10% auto; /* 15% from the top and centered */
+    padding: 20px;
+    border: 1px solid #888;
+    width: 60%;
+    border-radius: 0.75%;
+  }
+  
+  /* The Close Button */
+  .close-button {
+    color: #aaa;
+    float: right;
+    font-size: 28px;
+    font-weight: bold;
+  }
+  
+  .close-button:hover,
+  .close-button:focus {
+    color: black;
+    text-decoration: none;
+    cursor: pointer;
+  }
+  
+  .product-description {
+    height: 30px;
+  }
+
+  #adminBtn {
+    position: absolute;
+    margin-top: 5px;
+    margin-right: 15px;
+    right: 0;
+  }
