--- internationalcandyemporium/src/main/resources/templates/admin.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Admin</title>
-</head>
-<body>
-  Admin
-  <a href="/admin/addProduct">Add Product</a>
-  <a>Edit Product</a>
-</body>
-</html>
