+++ internationalcandyemporium/src/main/resources/templates/login.html
@@ -10,7 +10,7 @@
                 <a href="/profile/">Profile</a>
                 <a href="/cart/">View Cart</a>
                 <a href="/login/">Sign in</a>
-                <a href="/logout/">Log Out</a>
+                <a href="/login/">Log Out</a>
             </div>
         </div>
         <h1>International Candy Emporium</h1>
